refactor: extract input mouse helpers
