+++ src/app/input/mod.rs
@@ -1,13 +1,10 @@
 //! Input handling — translates crossterm key/mouse events into state mutations.
 
-use bytes::Bytes;
 use crossterm::event::{KeyCode, KeyEvent, MouseButton, MouseEvent, MouseEventKind};
 
 use crate::input::TerminalKey;
 use ratatui::layout::{Direction, Rect};
-use tracing::warn;
 
-use crate::layout::{PaneInfo, SplitBorder};
 use crate::selection::Selection;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -28,6 +25,7 @@ const WORKSPACE_DRAG_THRESHOLD: u16 = 1;
 const TAB_DRAG_THRESHOLD: u16 = 1;
 
 mod modal;
+mod mouse;
 mod navigate;
 mod overlays;
 mod settings;
@@ -37,6 +35,7 @@ mod terminal;
 #[cfg(test)]
 use self::{
     modal::open_rename_active_tab,
+    mouse::wheel_routing,
     navigate::{execute_navigate_action, handle_navigate_reserved_key, NavigateAction},
     settings::{open_settings, update_settings_state},
 };
@@ -883,118 +882,6 @@ impl AppState {
         crate::ui::confirm_close_popup_rect(self.view.terminal_area).unwrap_or_default()
     }
 
-    /// Find which workspace index a sidebar row belongs to (two-section layout).
-    fn tab_at(&self, col: u16, row: u16) -> Option<usize> {
-        self.view
-            .tab_hit_areas
-            .iter()
-            .enumerate()
-            .find_map(|(idx, area)| {
-                (area.width > 0
-                    && row >= area.y
-                    && row < area.y + area.height
-                    && col >= area.x
-                    && col < area.x + area.width)
-                    .then_some(idx)
-            })
-    }
-
-    fn on_tab_bar(&self, col: u16, row: u16) -> bool {
-        let area = self.view.tab_bar_rect;
-        area.width > 0
-            && row >= area.y
-            && row < area.y + area.height
-            && col >= area.x
-            && col < area.x + area.width
-    }
-
-    fn on_tab_scroll_left_button(&self, col: u16, row: u16) -> bool {
-        let area = self.view.tab_scroll_left_hit_area;
-        area.width > 0
-            && row >= area.y
-            && row < area.y + area.height
-            && col >= area.x
-            && col < area.x + area.width
-    }
-
-    fn on_tab_scroll_right_button(&self, col: u16, row: u16) -> bool {
-        let area = self.view.tab_scroll_right_hit_area;
-        area.width > 0
-            && row >= area.y
-            && row < area.y + area.height
-            && col >= area.x
-            && col < area.x + area.width
-    }
-
-    fn tab_drop_index_at(&self, col: u16, row: u16) -> Option<usize> {
-        if !self.on_tab_bar(col, row) {
-            return None;
-        }
-
-        let visible_tabs: Vec<_> = self
-            .view
-            .tab_hit_areas
-            .iter()
-            .enumerate()
-            .filter(|(_, rect)| rect.width > 0)
-            .collect();
-        let (first_idx, first_rect) = *visible_tabs.first()?;
-        let (last_idx, last_rect) = *visible_tabs.last()?;
-
-        if self.on_tab_scroll_left_button(col, row) {
-            return Some(0);
-        }
-        if self.on_tab_scroll_right_button(col, row) {
-            return self
-                .active
-                .and_then(|idx| self.workspaces.get(idx))
-                .map(|ws| ws.tabs.len());
-        }
-
-        let left_edge = if first_idx == 0 {
-            first_rect.x
-        } else {
-            self.view.tab_scroll_left_hit_area.x + self.view.tab_scroll_left_hit_area.width
-        };
-        let right_edge = if self
-            .active
-            .and_then(|idx| self.workspaces.get(idx))
-            .is_some_and(|ws| last_idx + 1 >= ws.tabs.len())
-        {
-            last_rect.x + last_rect.width
-        } else {
-            self.view.tab_scroll_right_hit_area.x.saturating_sub(1)
-        };
-
-        if col <= left_edge {
-            return Some(first_idx);
-        }
-        if col >= right_edge {
-            return Some(last_idx + 1);
-        }
-
-        for (idx, rect) in visible_tabs {
-            let midpoint = rect.x + rect.width / 2;
-            if col < midpoint {
-                return Some(idx);
-            }
-            if col < rect.x + rect.width {
-                return Some(idx + 1);
-            }
-        }
-
-        Some(last_idx + 1)
-    }
-
-    fn on_new_tab_button(&self, col: u16, row: u16) -> bool {
-        let area = self.view.new_tab_hit_area;
-        area.width > 0
-            && row >= area.y
-            && row < area.y + area.height
-            && col >= area.x
-            && col < area.x + area.width
-    }
-
     fn context_menu_item_at(&self, col: u16, row: u16) -> Option<usize> {
         let menu_rect = self.context_menu_rect()?;
         let inner_x = menu_rect.x + 1;
@@ -1015,83 +902,6 @@ impl AppState {
         } else {
             None
         }
-    }
-
-    fn find_border_at(&self, col: u16, row: u16) -> Option<&SplitBorder> {
-        self.view.split_borders.iter().find(|b| match b.direction {
-            Direction::Horizontal => {
-                (col as i32 - b.pos as i32).unsigned_abs() <= 1
-                    && row >= b.area.y
-                    && row < b.area.y + b.area.height
-            }
-            Direction::Vertical => {
-                (row as i32 - b.pos as i32).unsigned_abs() <= 1
-                    && col >= b.area.x
-                    && col < b.area.x + b.area.width
-            }
-        })
-    }
-
-    fn pane_at(&self, col: u16, row: u16) -> Option<&PaneInfo> {
-        self.view.pane_infos.iter().find(|p| {
-            col >= p.inner_rect.x
-                && col < p.inner_rect.x + p.inner_rect.width
-                && row >= p.inner_rect.y
-                && row < p.inner_rect.y + p.inner_rect.height
-        })
-    }
-
-    fn pane_mouse_target(&self, col: u16, row: u16) -> Option<&PaneInfo> {
-        self.pane_at(col, row)
-            .or_else(|| self.pane_frame_at(col, row))
-    }
-
-    fn pane_info_by_id(&self, pane_id: crate::layout::PaneId) -> Option<&PaneInfo> {
-        self.view.pane_infos.iter().find(|info| info.id == pane_id)
-    }
-
-    fn pane_frame_at(&self, col: u16, row: u16) -> Option<&PaneInfo> {
-        self.view.pane_infos.iter().find(|p| {
-            col >= p.rect.x
-                && col < p.rect.x + p.rect.width
-                && row >= p.rect.y
-                && row < p.rect.y + p.rect.height
-        })
-    }
-
-    fn focus_pane(&mut self, pane_id: crate::layout::PaneId) {
-        if let Some(ws) = self.active.and_then(|i| self.workspaces.get_mut(i)) {
-            if ws.layout.focused() != pane_id {
-                ws.layout.focus_pane(pane_id);
-                self.mark_session_dirty();
-            }
-        }
-    }
-
-    fn scroll_pane_up(&self, pane_id: crate::layout::PaneId, lines: usize) {
-        if let Some(ws) = self.active.and_then(|i| self.workspaces.get(i)) {
-            if let Some(rt) = ws.runtimes.get(&pane_id) {
-                rt.scroll_up(lines);
-            }
-        }
-    }
-
-    fn scroll_pane_down(&self, pane_id: crate::layout::PaneId, lines: usize) {
-        if let Some(ws) = self.active.and_then(|i| self.workspaces.get(i)) {
-            if let Some(rt) = ws.runtimes.get(&pane_id) {
-                rt.scroll_down(lines);
-            }
-        }
-    }
-
-    fn pane_scroll_metrics(
-        &self,
-        pane_id: crate::layout::PaneId,
-    ) -> Option<crate::pane::ScrollMetrics> {
-        self.active
-            .and_then(|i| self.workspaces.get(i))
-            .and_then(|ws| ws.runtime(pane_id))
-            .and_then(crate::pane::PaneRuntime::scroll_metrics)
     }
 
     fn update_selection_cursor(
@@ -1155,175 +965,6 @@ impl AppState {
         }
         self.update_selection_cursor(pane_id, mouse.column, mouse.row);
         true
-    }
-
-    fn handle_terminal_wheel(&mut self, mouse: MouseEvent) {
-        const LINES_PER_NOTCH: usize = 3;
-
-        if let Some(info) = self.pane_at(mouse.column, mouse.row).cloned() {
-            self.focus_pane(info.id);
-            if self.forward_pane_wheel(&info, mouse) {
-                return;
-            }
-            match mouse.kind {
-                MouseEventKind::ScrollUp => self.scroll_pane_up(info.id, LINES_PER_NOTCH),
-                MouseEventKind::ScrollDown => self.scroll_pane_down(info.id, LINES_PER_NOTCH),
-                _ => {}
-            }
-            return;
-        }
-
-        if let Some(info) = self.pane_frame_at(mouse.column, mouse.row).cloned() {
-            self.focus_pane(info.id);
-            match mouse.kind {
-                MouseEventKind::ScrollUp => self.scroll_pane_up(info.id, LINES_PER_NOTCH),
-                MouseEventKind::ScrollDown => self.scroll_pane_down(info.id, LINES_PER_NOTCH),
-                _ => {}
-            }
-            return;
-        }
-
-        if let Some(ws) = self.active.and_then(|i| self.workspaces.get(i)) {
-            if let Some(rt) = ws.focused_runtime() {
-                match mouse.kind {
-                    MouseEventKind::ScrollUp => rt.scroll_up(LINES_PER_NOTCH),
-                    MouseEventKind::ScrollDown => rt.scroll_down(LINES_PER_NOTCH),
-                    _ => {}
-                }
-            }
-        }
-    }
-
-    fn forward_pane_mouse_button(&self, info: &PaneInfo, mouse: MouseEvent) -> bool {
-        let Some(ws) = self.active.and_then(|i| self.workspaces.get(i)) else {
-            return false;
-        };
-        let Some(rt) = ws.runtimes.get(&info.id) else {
-            return false;
-        };
-        let column = mouse.column.saturating_sub(info.inner_rect.x);
-        let row = mouse.row.saturating_sub(info.inner_rect.y);
-        let Some(bytes) = rt.encode_mouse_button(mouse.kind, column, row, mouse.modifiers) else {
-            return false;
-        };
-        rt.scroll_reset();
-        if let Err(err) = rt.try_send_bytes(Bytes::from(bytes)) {
-            warn!(pane = info.id.raw(), err = %err, kind = ?mouse.kind, "failed to forward mouse button event");
-        }
-        true
-    }
-
-    fn forward_pane_wheel(&self, info: &PaneInfo, mouse: MouseEvent) -> bool {
-        let Some(ws) = self.active.and_then(|i| self.workspaces.get(i)) else {
-            return false;
-        };
-        let Some(rt) = ws.runtimes.get(&info.id) else {
-            return false;
-        };
-        match rt.wheel_routing() {
-            Some(crate::pane::WheelRouting::HostScroll) | None => false,
-            Some(crate::pane::WheelRouting::MouseReport) => {
-                rt.scroll_reset();
-                let column = mouse.column.saturating_sub(info.inner_rect.x);
-                let row = mouse.row.saturating_sub(info.inner_rect.y);
-                let Some(bytes) = rt.encode_mouse_wheel(mouse.kind, column, row, mouse.modifiers)
-                else {
-                    warn!(pane = info.id.raw(), kind = ?mouse.kind, "failed to encode mouse wheel event");
-                    return true;
-                };
-                if let Err(err) = rt.try_send_bytes(Bytes::from(bytes)) {
-                    warn!(pane = info.id.raw(), err = %err, "failed to forward mouse wheel event");
-                }
-                true
-            }
-            Some(crate::pane::WheelRouting::AlternateScroll) => {
-                rt.scroll_reset();
-                let Some(bytes) = rt.encode_alternate_scroll(mouse.kind) else {
-                    return true;
-                };
-                if let Err(err) = rt.try_send_bytes(Bytes::from(bytes)) {
-                    warn!(pane = info.id.raw(), err = %err, "failed to forward alternate-scroll key");
-                }
-                true
-            }
-        }
-    }
-
-    fn set_pane_scroll_offset(&self, pane_id: crate::layout::PaneId, offset_from_bottom: usize) {
-        if let Some(ws) = self.active.and_then(|i| self.workspaces.get(i)) {
-            if let Some(rt) = ws.runtimes.get(&pane_id) {
-                rt.set_scroll_offset_from_bottom(offset_from_bottom);
-            }
-        }
-    }
-
-    fn scrollbar_target_at(
-        &self,
-        col: u16,
-        row: u16,
-    ) -> Option<(crate::layout::PaneId, ScrollbarClickTarget)> {
-        let ws = self.active.and_then(|i| self.workspaces.get(i))?;
-        let info = self.view.pane_infos.iter().find(|info| {
-            crate::ui::pane_scrollbar_rect(info).is_some_and(|track| {
-                col >= track.x
-                    && col < track.x + track.width
-                    && row >= track.y
-                    && row < track.y + track.height
-            })
-        })?;
-        let rt = ws.runtimes.get(&info.id)?;
-        let metrics = rt.scroll_metrics()?;
-        if metrics.max_offset_from_bottom == 0 {
-            return None;
-        }
-        let track = crate::ui::pane_scrollbar_rect(info)?;
-        if let Some(grab_row_offset) = crate::ui::scrollbar_thumb_grab_offset(metrics, track, row) {
-            Some((info.id, ScrollbarClickTarget::Thumb { grab_row_offset }))
-        } else {
-            Some((
-                info.id,
-                ScrollbarClickTarget::Track {
-                    offset_from_bottom: crate::ui::scrollbar_offset_from_row(metrics, track, row),
-                },
-            ))
-        }
-    }
-
-    fn scrollbar_offset_for_pane_row(
-        &self,
-        pane_id: crate::layout::PaneId,
-        row: u16,
-        grab_row_offset: u16,
-    ) -> Option<usize> {
-        let ws = self.active.and_then(|i| self.workspaces.get(i))?;
-        let info = self
-            .view
-            .pane_infos
-            .iter()
-            .find(|info| info.id == pane_id)?;
-        let track = crate::ui::pane_scrollbar_rect(info)?;
-        let rt = ws.runtimes.get(&pane_id)?;
-        let metrics = rt.scroll_metrics()?;
-        if metrics.max_offset_from_bottom == 0 {
-            return None;
-        }
-        Some(crate::ui::scrollbar_offset_from_drag_row(
-            metrics,
-            track,
-            row,
-            grab_row_offset,
-        ))
-    }
-}
-
-#[cfg(test)]
-fn wheel_routing(input_state: crate::pane::InputState) -> WheelRouting {
-    if input_state.mouse_protocol_mode.reporting_enabled() {
-        WheelRouting::MouseReport
-    } else if input_state.alternate_screen && input_state.mouse_alternate_scroll {
-        WheelRouting::AlternateScroll
-    } else {
-        WheelRouting::HostScroll
     }
 }
